Added browser-based semi-auto-generated tests
A `MochaLoader.ts` and `index.html` are copied to package `test/` dirs
now. The `index.html` file is auto-generated from a mustache template
that takes in the package name, dependencies, and (from the package's
test's `tsconfig.json`) test names.

The `MochaLoader` class reads in the scripts as they come in and buckets
them in the appropriate `describe()` group in mocha.